drivers:misc:stml0xx: Extend WAKE_SENSORS array to 24 bits -- L0
Extended the wake_sensor array from 2 to 3 bytes (24 bits).
Unused flags are listed as M_UNUSED (commented out).

Reviewed-on: http://gerrit.mot.com/731651
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ drivers/misc/stml0xx/stml0xx_core.c
@@ -65,7 +65,7 @@ unsigned short stml0xx_g_mag_delay;
 unsigned short stml0xx_g_gyro_delay;
 unsigned short stml0xx_g_baro_delay;
 unsigned long stml0xx_g_nonwake_sensor_state;
-unsigned short stml0xx_g_wake_sensor_state;
+unsigned long stml0xx_g_wake_sensor_state;
 unsigned short stml0xx_g_algo_state;
 unsigned char stml0xx_g_motion_dur;
 unsigned char stml0xx_g_zmotion_dur;
